refactor: match Game struct, fix namespaces
- replaced CoverImagePath with CoverPath across services, views, and seed data to match Game struct
- moved MainWindowViewModel to PsConsoleLauncher.ViewModels namespace
- fixed MainWindow.xaml class reference to PsConsoleLauncher.views.MainWindow
- updated app.xaml to PsConsoleLauncher.App with Views/MainWindow startup
- updated games.json to use CoverPath key for consistency with model
