+++ Views/MainWindow.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="ConsoleMenu.MainWindow"
+<Window x:Class="PsConsoleLauncher.views.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="Console Menu" Height="450" Width="800"
