+++ app.xaml
@@ -1,9 +1,6 @@
-﻿<Application x:Class="ConsoleMenu.App"
+﻿<Application x:Class="PsConsoleLauncher.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="MainWindow.xaml">
-    <Application.Resources>
-    </Application.Resources>
+             StartupUri="Views/MainWindow.xaml">
+    <Application.Resources/>
 </Application>
-
-
