+++ games.json
@@ -3,6 +3,6 @@
 		"Title": "Notepad",
 		"Platform": "Windows",
 		"ExecutablePath": "notepad.exe",
-		"CoverImagePath": "Themes/Default/assets/notepad.png"
+		"CoverPath": "Themes/Default/assets/notepad.png"
 	}
 ]
